Factor out strings and common tasks.
