+++ src/uwtools/cli.py
@@ -7,7 +7,6 @@ from argparse import ArgumentParser as Parser
 from argparse import HelpFormatter
 from argparse import _ArgumentGroup as Group
 from argparse import _SubParsersAction as Subparsers
-from dataclasses import dataclass
 from functools import partial
 from pathlib import Path
 from typing import Any, Callable, Dict, List, NoReturn, Tuple
@@ -22,7 +21,8 @@ import uwtools.config.jinja2
 import uwtools.rocoto
 from uwtools.exceptions import UWConfigRealizeError, UWError, UWTemplateRenderError
 from uwtools.logging import log, setup_logging
-from uwtools.utils.file import FORMAT, get_file_format
+from uwtools.strings import FORMAT, STR
+from uwtools.utils.file import get_file_format
 
 FORMATS = FORMAT.extensions()
 TITLE_REQ_ARG = "Required arguments"
@@ -916,53 +916,3 @@ def _switch(arg: str) -> str:
     :return: The long-form switch.
     """
     return "--%s" % arg.replace("_", "-")
-
-
-@dataclass(frozen=True)
-class STR:
-    """
-    A lookup map for CLI-related strings.
-    """
-
-    action: str = "action"
-    batch: str = "batch"
-    cfgfile: str = "config_file"
-    chgrescube: str = "chgres_cube"
-    compare: str = "compare"
-    config: str = "config"
-    cycle: str = "cycle"
-    dryrun: str = "dry_run"
-    env: str = "env"
-    file1fmt: str = "file_1_format"
-    file1path: str = "file_1_path"
-    file2fmt: str = "file_2_format"
-    file2path: str = "file_2_path"
-    fv3: str = "fv3"
-    graphfile: str = "graph_file"
-    help: str = "help"
-    infile: str = "input_file"
-    infmt: str = "input_format"
-    keyvalpairs: str = "key_eq_val_pairs"
-    mode: str = "mode"
-    model: str = "model"
-    outfile: str = "output_file"
-    outfmt: str = "output_format"
-    partial: str = "partial"
-    quiet: str = "quiet"
-    realize: str = "realize"
-    render: str = "render"
-    rocoto: str = "rocoto"
-    run: str = "run"
-    schemafile: str = "schema_file"
-    sfcclimogen: str = "sfc_climo_gen"
-    suppfiles: str = "supplemental_files"
-    task: str = "task"
-    tasks: str = "tasks"
-    template: str = "template"
-    total: str = "total"
-    translate: str = "translate"
-    validate: str = "validate"
-    valsfile: str = "values_file"
-    valsfmt: str = "values_format"
-    valsneeded: str = "values_needed"
-    verbose: str = "verbose"
