+++ src/uwtools/drivers/chgres_cube.py
@@ -12,6 +12,7 @@ from iotaa import asset, dryrun, task, tasks
 
 from uwtools.config.formats.nml import NMLConfig
 from uwtools.drivers.driver import Driver
+from uwtools.strings import STR
 from uwtools.utils.processing import execute
 from uwtools.utils.tasks import file
 
@@ -20,6 +21,8 @@ class ChgresCube(Driver):
     """
     A driver for chgres_cube.
     """
+
+    _driver_name = STR.chgrescube
 
     def __init__(
         self, config_file: Path, cycle: datetime, dry_run: bool = False, batch: bool = False
